+++ R/expressions.R
@@ -35,6 +35,9 @@ setMethod("as.vector", "CrunchExpr", function (x, mode) {
     return(columnParser(out$metadata$out$type, mode)(out$data$out, variable))
 })
 
+setMethod("toVariable", "CrunchExpr", function (x, ...) {
+    structure(list(expr=zcl(x), ...), class="VariableDefinition")
+})
 
 ## "Ops" for Crunch Variables
 ##